### 2025-11-20 #9479 @editorialbot
[PRE REVIEW]: tools4RDF: A python toolkit for working with RDF data
**Submitting author:** @srmnitc (<a href="http://orcid.org/0000-0002-6776-1213">Sarath Menon</a>)
**Repository:** https://github.com/OCDO/tools4RDF
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.6
