### 2025-11-20 #9479 @crvernon
@srmnitc - I am going to take this one.  While I am finding you reviewers, please address the missing or invalid DOI listed [above](https://github.com/openjournals/joss-reviews/issues/9479#issuecomment-3558079761).
