### 2025-11-20 #9479 @r0hin
> 👋 @r0hin - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Yes, of course.
