### 2025-11-20 c4c5aeb
fix paper.bib DOIs
