+++ joss/paper.bib
@@ -39,11 +39,10 @@ year = {2025}
 
 
 
-@article{cuddihy_semtk_nodate,
-	title = {{SemTK}: {A} {Semantics} {Toolkit} for {User}-friendly {SPARQL} {Generation} and {Semantic} {Data} {Management}},
-	language = {en},
-        url={https://github.com/ge-semtk/semtk},
-	author = {Cuddihy, Paul and McHugh, Justin and Williams, Jenny Weisenberg and Mulwad, Varish and Aggour, Kareem},
+@software{cuddihy_semtk_nodate,
+  title        = {{SemTK: A Semantics Toolkit for User-friendly SPARQL Generation and Semantic Data Management}},
+  url          = {https://github.com/ge-semtk/semtk},
+  note         = {Accessed: 2025-11-20}
 }
 
 
@@ -59,20 +58,21 @@ year = {2025}
 
 @misc{w3c-rdf11,
   author       = {{W3C RDF Working Group}},
-  title        = {{RDF 1.1 Concepts and Abstract Syntax}},
+  title        = {RDF 1.1 Concepts and Abstract Syntax},
   year         = {2014},
-  howpublished = {\url{https://www.w3.org/TR/rdf11-concepts/}},
+  institution  = {World Wide Web Consortium},
+  url          = {https://www.w3.org/TR/rdf11-concepts/},
   note         = {W3C Recommendation}
 }
 
 @misc{w3c-sparql11,
   author       = {{W3C SPARQL Working Group}},
-  title        = {{SPARQL 1.1 Query Language}},
+  title        = {SPARQL 1.1 Query Language},
   year         = {2013},
-  howpublished = {\url{https://www.w3.org/TR/sparql11-query/}},
+  institution  = {World Wide Web Consortium},
+  url          = {https://www.w3.org/TR/sparql11-query/},
   note         = {W3C Recommendation}
 }
-
 
 @inproceedings{gutierrez_knowledge_2020,
 	address = {Virtual Event Ireland},
@@ -132,6 +132,7 @@ booktitle = {Proceedings of the 6th International The Semantic Web and 2nd Asian
 pages = {722–735},
 numpages = {14},
 location = {Busan, Korea},
+doi = {10.1007/978-3-540-76298-0_52}
 }
 
 @article{gansner2000open,
@@ -142,7 +143,7 @@ location = {Busan, Korea},
   number={11},
   pages={1203--1233},
   year={2000},
-  doi = {10.1002/1097-024X(200009)30:11\%3C1203::AID-SPE338\%3E3.0.CO;2-N},
+  doi = {{10.1002/1097-024X(200009)30:11%3C1203::AID-SPE338%3E3.0.CO;2-N}},
   publisher={Wiley Online Library}
 }  
 
