### 2025-11-20 59a0810
Fix broken DOI in paper.bib
