+++ joss/paper.bib
@@ -143,7 +143,7 @@ doi = {10.1007/978-3-540-76298-0_52}
   number={11},
   pages={1203--1233},
   year={2000},
-  doi = {{10.1002/1097-024X(200009)30:11%3C1203::AID-SPE338%3E3.0.CO;2-N}},
+  doi = {10.1002/1097-024X(200009)30:11<1203::AID-SPE338>3.0.CO;2-N},
   publisher={Wiley Online Library}
 }  
 
