### 2025-11-20 9b1b860
fix the wrong resolution in parser
