### 2025-11-20 4250628
fix for some subclasses not being added
