### 2025-11-21 1485db4
fix some unnecessary braces
