### 2025-11-21 21f1cdb
clean up some operators
