### 2025-11-21 845a4f9
fix for condition resetting
