+++ tests/test_network.py
@@ -1,15 +1,475 @@
 import pytest
 from tools4rdf.network.ontology import read_ontology
-from rdflib import Graph
+from tools4rdf.network.network import (
+    Network,
+    OntologyNetwork,
+    OntologyNetworkBase,
+    _replace_name,
+    _strip_name,
+)
+from tools4rdf.network.parser import OntoParser
+from tools4rdf.network.term import OntoTerm
+from rdflib import Graph, Namespace, RDF, RDFS, OWL, URIRef
+
 
 def test_network():
     onto = read_ontology()
     kg = Graph()
     kg.parse("tests/triples", format="turtle")
-    df = onto.query(kg, onto.terms.cmso.AtomicScaleSample, [onto.terms.cmso.hasSpaceGroupSymbol, onto.terms.cmso.hasNumberOfAtoms==4])
+    df = onto.query(
+        kg,
+        onto.terms.cmso.AtomicScaleSample,
+        [onto.terms.cmso.hasSpaceGroupSymbol, onto.terms.cmso.hasNumberOfAtoms == 4],
+    )
     assert len(df) == 14
+
 
 def test_owlThing():
     onto = read_ontology()
-    query = (onto.create_query(onto.terms.cmso.AtomicScaleSample, [[onto.terms.cmso.CrystalStructure, onto.terms.cmso.hasAltName]]))
+    query = onto.create_query(
+        onto.terms.cmso.AtomicScaleSample,
+        [[onto.terms.cmso.CrystalStructure, onto.terms.cmso.hasAltName]],
+    )
     assert "CrystalStructure_hasAltNamevalue" in query
+
+
+@pytest.fixture
+def simple_onto():
+    g = Graph()
+    ex = Namespace("http://example.org/ontology#")
+    g.bind("ex", ex)
+    g.add((ex.Pizza, RDF.type, OWL.Class))
+    g.add((ex.Food, RDF.type, OWL.Class))
+    g.add((ex.Person, RDF.type, OWL.Class))
+    g.add((ex.Pizza, RDFS.subClassOf, ex.Food))
+    g.add((ex.hasTopping, RDF.type, OWL.ObjectProperty))
+    g.add((ex.hasTopping, RDFS.domain, ex.Pizza))
+    g.add((ex.hasTopping, RDFS.range, ex.Food))
+    g.add((ex.hasPrice, RDF.type, OWL.DatatypeProperty))
+    g.add((ex.hasPrice, RDFS.domain, ex.Pizza))
+    return OntoParser(g)
+
+
+@pytest.fixture
+def network_fixture(simple_onto):
+    return Network(simple_onto)
+
+
+def test_replace_name():
+    assert _replace_name("ex:Pizza") == "ex.Pizza"
+    assert _replace_name("simple") == "simple"
+
+
+def test_strip_name():
+    assert _strip_name("ex:Pizza") == "Pizza"
+    assert _strip_name("Pizza") == "Pizza"
+
+
+def test_network_init(simple_onto):
+    network = Network(simple_onto)
+    assert network.terms is not None
+    assert network.g is not None
+    assert network.namespaces is not None
+    assert network.extra_namespaces is not None
+
+
+def test_network_draw(network_fixture):
+    dot = network_fixture.draw()
+    assert dot is not None
+
+
+def test_network_draw_custom_style(network_fixture):
+    style = {
+        "class": {"shape": "circle"},
+        "object_property": {"shape": "diamond"},
+        "data_property": {"shape": "box"},
+        "literal": {"shape": "parallelogram"},
+    }
+    dot = network_fixture.draw(styledict=style)
+    assert dot is not None
+
+
+def test_get_shortest_path_simple(network_fixture):
[…]
