### 2025-12-01 b04a1ab
add logo to README
