+++ README.md
@@ -1,4 +1,5 @@
-## 🧰 tools4RDF
+
+![logo](docs/source/_static/logo.jpg "logo")
 
 **tools4RDF** is a Python toolkit for working with RDF data, SPARQL queries, and semantic networks. 
 
