### 2025-12-01 9abc9c4
Update Sphinx config path in ReadTheDocs YAML
