### 2025-12-01 525f3d6
Update Jupyter Book Sphinx config path
