+++ .readthedocs.yml
@@ -7,7 +7,7 @@ build:
   jobs:
     pre_build:
       # Generate the Sphinx configuration for this Jupyter Book so it builds.
-      - jupyter-book config sphinx docs/
+      - "jupyter-book config sphinx ."
 
 conda:
   environment: environment-docs.yml
