### 2025-12-01 3c626df
update logo to PNG
