### 2025-12-09 #9482 @r0hin
@srmnitc this project looks great. I put my suggestions in the review checklist, but here are them for completeness:

- Since you mention SemTK as a closely related toolkit, please add an appropriate citation (e.g. the SemTK paper or project URL) to the reference list.
- The documentation should briefly mention the target audience.
- It might be worth including a fallback source incase the foaf xmlns link is down on the main ipynb as it was for me -- up to you on this.

Everything else looks good, and I was indeed able to reproduce the project's functionality. I made a PR to address points 2 & 3 for you.
