### 2025-12-11 0202f1e
Update SemTK entry in bibliography
