+++ joss/paper.bib
@@ -38,13 +38,15 @@ year = {2025}
 }
 
 
-
-@software{cuddihy_semtk_nodate,
-  title        = {{SemTK: A Semantics Toolkit for User-friendly SPARQL Generation and Semantic Data Management}},
-  url          = {https://github.com/ge-semtk/semtk},
-  note         = {Accessed: 2025-11-20}
+@misc{cuddihy_semtk_nodate,
+      title={SemTK: An Ontology-first, Open Source Semantic Toolkit for Managing and Querying Knowledge Graphs}, 
+      author={Paul Cuddihy and Justin McHugh and Jenny Weisenberg Williams and Varish Mulwad and Kareem S. Aggour},
+      year={2018},
+      eprint={1710.11531},
+      archivePrefix={arXiv},
+      primaryClass={cs.AI},
+      url={https://arxiv.org/abs/1710.11531}, 
 }
-
 
 @InProceedings{ mckinney-proc-scipy-2010,
   author    = { {W}es {M}c{K}inney },
