### 2025-12-11 4e316b3
Update citation for SemTK reference in paper.bib
