+++ joss/paper.bib
@@ -38,7 +38,7 @@ year = {2025}
 }
 
 
-@misc{cuddihy_semtk_nodate,
+@misc{Cuddihy2018,
       title={SemTK: An Ontology-first, Open Source Semantic Toolkit for Managing and Querying Knowledge Graphs}, 
       author={Paul Cuddihy and Justin McHugh and Jenny Weisenberg Williams and Varish Mulwad and Kareem S. Aggour},
       year={2018},
