### 2025-12-11 47e0e08
Add SemTK citation in the Statement of Need
Updated references in the Statement of Need section to include SemTK citation.
