+++ joss/paper.md
@@ -48,7 +48,7 @@ tools4RDF is a lightweight Python framework designed to simplify working with RD
 
 # Statement of need
 
-Knowledge graphs [@gutierrez_knowledge_2020], coupled with formal ontologies, offer a promising technical approach for implementing the FAIR Guiding Principles [@wilkinson_fair_2016] for research data management. They support semantic interoperability, especially for aggregation and description of heterogeneous data [@vogt_suggestions_2025]. Accessing and querying these graphs typically relies on SPARQL [@w3c-sparql11], the standard query language for RDF data [@w3c-sparql11]. However, despite its potential, SPARQL remains difficult to use for many domain scientists due to unfamiliar syntax, steep learning curve, and the absence of a user-friendly tool stack [@Khamparia2017]. Libraries such as RDFLib [@Krech2025] offer core functionality for RDF parsing and graph storage but require users to write SPARQL queries and manipulate low-level graph objects. Tools like Protégé [@Musen2015] and [SemTK](https://github.com/ge-semtk/semtk) provide graphical interfaces for ontology exploration, but they focus on GUI-based interaction and lack integration with the Python ecosystem commonly used in scientific computing.
+Knowledge graphs [@gutierrez_knowledge_2020], coupled with formal ontologies, offer a promising technical approach for implementing the FAIR Guiding Principles [@wilkinson_fair_2016] for research data management. They support semantic interoperability, especially for aggregation and description of heterogeneous data [@vogt_suggestions_2025]. Accessing and querying these graphs typically relies on SPARQL [@w3c-sparql11], the standard query language for RDF data [@w3c-sparql11]. However, despite its potential, SPARQL remains difficult to use for many domain scientists due to unfamiliar syntax, steep learning curve, and the absence of a user-friendly tool stack [@Khamparia2017]. Libraries such as RDFLib [@Krech2025] offer core functionality for RDF parsing and graph storage but require users to write SPARQL queries and manipulate low-level graph objects. Tools like Protégé [@Musen2015] and SemTK [@Cuddihy2018] provide graphical interfaces for ontology exploration, but they focus on GUI-based interaction and lack integration with the Python ecosystem commonly used in scientific computing.
 
 tools4RDF fills this gap by providing a Python-native interface for ontology exploration and query construction. By ingesting one or more ontologies and converting them into navigable Python classes, allowing users to navigate class hierarchies, properties, and relationships through autocompletion. This enables users to perform predicate-based graph traversal and generate SPARQL queries programmatically, without requiring manual SPARQL scripting. The framework tools4RDF supports a code-first workflow that allows domain scientists to explore semantic data in familiar environments such as Jupyter notebooks. Query results can be parsed directly into [Pandas](https://doi.org/10.5281/zenodo.3509134) [@mckinney-proc-scipy-2010] dataframes, facilitating integration with existing data processing pipelines. Beyond simplifying ontology interaction, this programmatic access to ontologies may also be useful for integrating semantic data into applications involving agentic AI or large language models. The tool supports the construction, maintenance, and use of complex knowledge graphs in computational research.
 
