### 2025-12-28 #9482 @fcdimitr
@srmnitc Thanks for the submission. This package is a useful ergonomic layer for ontology-guided querying of RDF datasets, especially for Python/Jupyter workflows.

I have a few questions & suggestions that I believe would strengthen the paper:
1. Could you list the subset of SPARQL features currently supported? Either in the paper or in the docs with a link from the paper?
2. Could you add a short paragraph on validation and correctness? For example, what checks ensure the generated query matches the intended semantics, and what is the path-selection/validation policy when multiple ontology-valid routes exist?
3. The manuscript mentions that "this programmatic access to ontologies may also be useful for integrating semantic data into applications involving agentic AI or large language models". Since an LLM could already emit SPARQL, what is the added value here? For example, is it constraining generation to ontology-valid predicates/paths, structured tool calls, reduced query error rate? I think a concrete example would benefit the paper.

Overall, the package and documentation look solid and well-organized. My suggestions focus on the paper and how to strengthen the framing of the contribution.
