+++ joss/paper.bib
@@ -1,3 +1,32 @@
+@misc{Emonet2025,
+      title={LLM-based SPARQL Query Generation from Natural Language over Federated Knowledge Graphs}, 
+      author={Vincent Emonet and Jerven Bolleman and Severine Duvaud and Tarcisio Mendes de Farias and Ana Claudia Sima},
+      year={2025},
+      eprint={2410.06062},
+      archivePrefix={arXiv},
+      primaryClass={cs.DB},
+      url={https://arxiv.org/abs/2410.06062}, 
+}
+
+@misc{Meyer2025,
+      title={Assessing SPARQL capabilities of Large Language Models}, 
+      author={Lars-Peter Meyer and Johannes Frey and Felix Brei and Natanael Arndt},
+      year={2025},
+      eprint={2409.05925},
+      archivePrefix={arXiv},
+      primaryClass={cs.DB},
+      url={https://arxiv.org/abs/2409.05925}, 
+}
+
+@InProceedings{Hagberg2008	,
+  author =       {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
+  title =        {Exploring Network Structure, Dynamics, and Function using NetworkX},
+  booktitle =   {Proceedings of the 7th Python in Science Conference},
+  pages =     {11 - 15},
+  address = {Pasadena, CA USA},
+  year =      {2008},
+  editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
+}
 
 @article{Khamparia2017,
 	title = {Comprehensive analysis of semantic web reasoners and tools: a survey},
