### 2026-01-20 acc26eb
add link to supported commands
