+++ joss/paper.bib
@@ -26,6 +26,8 @@
   address = {Pasadena, CA USA},
   year =      {2008},
   editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
+  doi = {10.25080/TCWV9851},
+  url = {https://doi.org/10.25080/TCWV9851}
 }
 
 @article{Khamparia2017,
