+++ joss/paper.bib
@@ -79,7 +79,7 @@ year = {2025}
       url={https://arxiv.org/abs/1710.11531}, 
 }
 
-@InProceedings{ mckinney-proc-scipy-2010,
+@InProceedings{mckinney-proc-scipy-2010,
   author    = { {W}es {M}c{K}inney },
   title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
   booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
