+++ joss/paper.bib
@@ -18,7 +18,7 @@
       url={https://arxiv.org/abs/2409.05925}, 
 }
 
-@InProceedings{Hagberg2008	,
+@InProceedings{Hagberg2008,
   author =       {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
   title =        {Exploring Network Structure, Dynamics, and Function using NetworkX},
   booktitle =   {Proceedings of the 7th Python in Science Conference},
@@ -162,7 +162,7 @@ isbn = {3540762973},
 publisher = {Springer-Verlag},
 address = {Berlin, Heidelberg},
 booktitle = {Proceedings of the 6th International The Semantic Web and 2nd Asian Conference on Asian Semantic Web Conference},
-pages = {722–735},
+pages = {722--735},
 numpages = {14},
 location = {Busan, Korea},
 doi = {10.1007/978-3-540-76298-0_52}
@@ -191,6 +191,6 @@ number = {3},
 issn = {0168-7433},
 doi = {10.1007/s10817-014-9305-1},
 journal = {J. Autom. Reason.},
-pages = {245–269},
+pages = {245--269},
 numpages = {25},
 }
