### 2026-01-20 ed00766
add DOIs for archive papers
