+++ joss/paper.bib
@@ -5,7 +5,8 @@
       eprint={2410.06062},
       archivePrefix={arXiv},
       primaryClass={cs.DB},
-      url={https://arxiv.org/abs/2410.06062}, 
+      url={https://arxiv.org/abs/2410.06062},
+	  doi={10.48550/arXiv.2410.06062},ß 
 }
 
 @misc{Meyer2025,
@@ -15,7 +16,8 @@
       eprint={2409.05925},
       archivePrefix={arXiv},
       primaryClass={cs.DB},
-      url={https://arxiv.org/abs/2409.05925}, 
+      url={https://arxiv.org/abs/2409.05925},
+	  doi={10.48550/arXiv.2409.05925},
 }
 
 @InProceedings{Hagberg2008,
