+++ joss/paper.bib
@@ -6,7 +6,7 @@
       archivePrefix={arXiv},
       primaryClass={cs.DB},
       url={https://arxiv.org/abs/2410.06062},
-	  doi={10.48550/arXiv.2410.06062},ß 
+      doi={10.48550/arXiv.2410.06062}
 }
 
 @misc{Meyer2025,
@@ -17,7 +17,7 @@
       archivePrefix={arXiv},
       primaryClass={cs.DB},
       url={https://arxiv.org/abs/2409.05925},
-	  doi={10.48550/arXiv.2409.05925},
+      doi={10.48550/arXiv.2409.05925}
 }
 
 @InProceedings{Hagberg2008,
