### 2026-01-20 #9482 @srmnitc
@fcdimitr Thank you for the review and for the helpful suggestions. Apologies for the delayed response due to the holiday period. We have addressed all three points in the revised paper.

1. We now explicitly list the supported SPARQL features in the [documentation](https://tools4rdf.readthedocs.io/en/latest/#supported-sparql-keywords), and added a reference to this section in the paper on line 101.
2. We added a paragraph on validation and correctness in the paper on lines 85-95.
3. We expanded the discussion on agentic AI and large language models, and how they could interface with tools4RDF, on lines 48–55. While we do not yet include a concrete example, we plan to add an LLM integration layer in a future release.

We hope these changes adequately address your suggestions. Thank you again for your constructive feedback.
