### 2026-01-25 6586920
Bump version: 0.3.6 → 0.3.7
