### 2026-01-25 993a2e3
Update title in CITATION.cff for clarity
