### 2026-01-25 ba5fa8f
Fix citation title formatting in CITATION.cff
