### 2026-01-25 #9482 @srmnitc
Version 0.3.8.  10.5281/zenodo.18370768

> 👋 [@srmnitc](https://github.com/srmnitc) - I need you to make the following changes in your paper:
>
>     * [x]  The "p" in "python" should be capitalized in the title[x]  LINE 78:  The link in "FOAF" is broken
>
>
> Next is just setting up the archive for your new release.
>
> We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.
>
> So here is what we have left to do:
>
>     * [x]
>       [x]   Confirm that the title of your paper in your paper.md file has no special formatting (e.g., backticks) as this will cause formatting issues when accepted.
>       [x]
>       [x]   Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.
>       [x]
>       [x]   Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).
>       [x]
>       [x]   Please respond with the DOI of the archived version here along with the version tag name released.
>
>
> I can then move forward with accepting the submission.

@crvernon Thanks a lot! I have fixed the p. The FOAF link was momentarily down, but seems to be working again now.

I have released version 0.3.8, and archived it on Zenodo with DOI:  10.5281/zenodo.18370768. Please let me know if anything more is needed.
