### 2026-01-26 #9482 @crvernon
🥳 Congratulations on your new publication @srmnitc ! Many thanks to @r0hin and @fcdimitr for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
